fix JEI, last couple changes to get it all working

+++ src/main/java/net/dries007/tfc/common/blockentities/BloomBlockEntity.java
@@ -89,6 +89,6 @@ public class BloomBlockEntity extends TFCBlockEntity
     public BlockState getState()
     {
         assert level != null;
-        return count <= 0 ? Blocks.AIR.defaultBlockState() : level.getBlockState(worldPosition).setValue(BloomBlock.LAYERS, Mth.clamp(count / maxCount * TOTAL_LAYERS, 1, 8));
+        return count <= 0 ? Blocks.AIR.defaultBlockState() : level.getBlockState(worldPosition).setValue(BloomBlock.LAYERS, (int) Mth.clamp((float) count / maxCount * TOTAL_LAYERS, 1, 8));
     }
 }
